+++ ml_tools/IO_tools.py
@@ -3,6 +3,8 @@ from ._core._IO_tools import (
     train_logger,
     save_list_strings,
     load_list_strings,
+    save_json,
+    load_json,
     compare_lists,
     info
 )
@@ -10,6 +12,8 @@ from ._core._IO_tools import (
 __all__ = [
     "custom_logger",
     "train_logger",
+    "save_json",
+    "load_json",
     "save_list_strings",
     "load_list_strings",
     "compare_lists"